added code to the seed files based on the models
